requirements.txt
hydra-core
hydra-colorlog

pytest~=7.3.1

torch~=2.0.1
torchvision~=0.15.2
lightning~=2.0.2
lightning-bolts

omegaconf~=2.3.0
matplotlib~=3.7.1
tqdm~=4.65.0